web_server/render/app/requirements.txt
Flask==1.1.2
itsdangerous==2.0.1
python-dotenv==0.19.2 